ai-engine/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.1
solana>=0.36.0
solders>=0.23.0
pydantic>=2.7.0
